refactor some functions into hooks and utils
